feat(search): no longer manually concatenate name
Hey! 

I just enabled this with a query rule on the index instead, now automatically, when there's a match with any of the concatenated names, it will be boosted as an optionalFilter. 

This means that the configuration doesn't need to be done on your side anymore. You should see the exact same results after merging this PR